refactor: split workers and unify native dependencies
