+++ ai_worker/src/application.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace ai_worker {
+
+int Run(int argc, char* argv[]);
+
+}  // namespace ai_worker
