+++ aipc-daemon/src/api.rs
@@ -21,9 +21,14 @@ use std::net::SocketAddr;
 use std::path::Path;
 use std::path::PathBuf;
 use std::time::Duration;
-use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWrite, AsyncWriteExt, SeekFrom};
+use tokio::io::{AsyncReadExt, AsyncSeekExt, SeekFrom};
 use tokio_util::io::ReaderStream;
 use tower_http::services::{ServeDir, ServeFile};
+
+mod ai;
+mod zip64;
+use ai::*;
+use zip64::write_zip64;
 
 #[derive(Clone)]
 struct AppState {
@@ -121,145 +126,6 @@ struct WebRtcOffer {
 
 async fn webrtc_status(State(state): State<AppState>) -> impl IntoResponse {
     Json(state.webrtc.status())
-}
-
-async fn ai_status(State(state): State<AppState>) -> impl IntoResponse {
-    Json(state.ai.status())
-}
-
-async fn ai_projects(State(state): State<AppState>) -> Result<impl IntoResponse, AppError> {
-    Ok(Json(state.ai.list_projects().await?))
-}
-
-async fn ai_project_create(
-    State(state): State<AppState>,
-    Json(document): Json<AiProjectDocument>,
-) -> Result<impl IntoResponse, AppError> {
-    let id = document.manifest.id.clone();
-    Ok((
-        StatusCode::CREATED,
-        Json(state.ai.put_project(&id, document).await?),
-    ))
-}
-
-async fn ai_project_get(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<impl IntoResponse, AppError> {
-    Ok(Json(state.ai.get_project(&id).await?))
-}
-
-async fn ai_project_put(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-    Json(document): Json<AiProjectDocument>,
-) -> Result<impl IntoResponse, AppError> {
-    Ok(Json(state.ai.put_project(&id, document).await?))
-}
-
-async fn ai_project_delete(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<StatusCode, AppError> {
-    state.ai.delete_project(&id).await?;
-    Ok(StatusCode::NO_CONTENT)
-}
-
-async fn ai_project_validate(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<impl IntoResponse, AppError> {
-    Ok(Json(state.ai.validate_project(&id).await?))
-}
-
-async fn ai_project_deploy(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<impl IntoResponse, AppError> {
-    Ok((StatusCode::ACCEPTED, Json(state.ai.deploy(&id).await?)))
-}
-
-async fn ai_models(State(state): State<AppState>) -> Result<impl IntoResponse, AppError> {
-    Ok(Json(state.ai.list_models().await?))
-}
-
-async fn ai_model_upload(
-    State(state): State<AppState>,
-    mut multipart: Multipart,
-) -> Result<impl IntoResponse, AppError> {
-    let mut name = None;
-    let mut data = None;
-    while let Some(field) = multipart.next_field().await? {
-        if field.name() == Some("file") {
-            name = field.file_name().map(ToOwned::to_owned);
-            data = Some(field.bytes().await?);
-        }
-    }
-    let name = name.ok_or_else(|| AppError::bad_request("multipart file name is required"))?;
-    let data = data.ok_or_else(|| AppError::bad_request("multipart file field is required"))?;
-    Ok((
-        StatusCode::CREATED,
-        Json(state.ai.put_model(&name, &data).await?),
-    ))
-}
-
-async fn ai_model_delete(
-    State(state): State<AppState>,
-    AxumPath(name): AxumPath<String>,
-) -> Result<StatusCode, AppError> {
-    state.ai.delete_model(&name).await?;
-    Ok(StatusCode::NO_CONTENT)
-}
-
-async fn ai_osd(State(state): State<AppState>) -> impl IntoResponse {
-    Json(json!({"mode": state.ai.status().osd_mode}))
-}
-
-#[derive(Deserialize)]
-struct OsdUpdate {
-    mode: OsdMode,
-}
-
-async fn ai_osd_update(
-    State(state): State<AppState>,
-    Json(update): Json<OsdUpdate>,
-) -> Result<impl IntoResponse, AppError> {
-    Ok(Json(
-        json!({"mode": state.ai.set_osd_mode(update.mode).await?}),
-    ))
-}
-
-async fn ai_events(
-    State(state): State<AppState>,
-) -> Sse<impl futures_util::Stream<Item = Result<Event, Infallible>>> {
-    let receiver = state.ai.subscribe_metadata();
-    let stream = stream::unfold(receiver, |mut receiver| async move {
-        loop {
-            match receiver.recv().await {
-                Ok(metadata) => {
-                    let item = Event::default()
-                        .event("detections")
-                        .json_data(&*metadata)
-                        .unwrap_or_else(|_| Event::default().event("serialization_error"));
-                    return Some((Ok(item), receiver));
-                }
-                Err(tokio::sync::broadcast::error::RecvError::Lagged(skipped)) => {
-                    return Some((
-                        Ok(Event::default()
-                            .event("lagged")
-                            .data(json!({"skipped": skipped}).to_string())),
-                        receiver,
-                    ));
-                }
-                Err(tokio::sync::broadcast::error::RecvError::Closed) => return None,
-            }
-        }
-    });
-    Sse::new(stream).keep_alive(
-        KeepAlive::new()
-            .interval(Duration::from_secs(10))
-            .text("keep-alive"),
-    )
 }
 
 async fn webrtc_create(
@@ -532,148 +398,6 @@ async fn recordings_export(
             format!("attachment; filename=\"{file_name}\""),
         )
         .body(Body::from_stream(ReaderStream::new(reader)))?)
-}
-
-struct ZipEntry {
-    name: Vec<u8>,
-    crc: u32,
-    size: u64,
-    offset: u64,
-}
-
-async fn write_zip64<W: AsyncWrite + Unpin>(
-    mut writer: W,
-    files: Vec<(String, PathBuf)>,
-) -> anyhow::Result<()> {
-    let mut entries = Vec::new();
-    let mut offset = 0_u64;
-    for (name, path) in files {
-        let safe_name: Vec<u8> = name
-            .bytes()
-            .map(|byte| {
-                if matches!(byte, b'/' | b'\\') {
-                    b'_'
-                } else {
-                    byte
-                }
-            })
-            .collect();
-        let local_offset = offset;
-        let mut header = Vec::new();
-        push_u32(&mut header, 0x04034b50);
-        push_u16(&mut header, 45);
-        push_u16(&mut header, 0x0008);
-        push_u16(&mut header, 0);
-        push_u16(&mut header, 0);
-        push_u16(&mut header, 0);
-        push_u32(&mut header, 0);
-        push_u32(&mut header, u32::MAX);
-        push_u32(&mut header, u32::MAX);
-        push_u16(&mut header, safe_name.len() as u16);
-        push_u16(&mut header, 20);
-        header.extend_from_slice(&safe_name);
-        push_u16(&mut header, 0x0001);
-        push_u16(&mut header, 16);
-        push_u64(&mut header, 0);
-        push_u64(&mut header, 0);
-        writer.write_all(&header).await?;
-        offset += header.len() as u64;
-        let mut file = tokio::fs::File::open(path).await?;
-        let mut buffer = vec![0_u8; 64 * 1024];
-        let mut hasher = crc32fast::Hasher::new();
-        let mut size = 0_u64;
-        loop {
-            let count = file.read(&mut buffer).await?;
-            if count == 0 {
-                break;
-            }
-            hasher.update(&buffer[..count]);
-            writer.write_all(&buffer[..count]).await?;
-            size += count as u64;
-            offset += count as u64;
-        }
-        let crc = hasher.finalize();
-        let mut descriptor = Vec::new();
-        push_u32(&mut descriptor, 0x08074b50);
-        push_u32(&mut descriptor, crc);
-        push_u64(&mut descriptor, size);
-        push_u64(&mut descriptor, size);
-        writer.write_all(&descriptor).await?;
-        offset += descriptor.len() as u64;
-        entries.push(ZipEntry {
-            name: safe_name,
-            crc,
-            size,
-            offset: local_offset,
-        });
-    }
-    let central_offset = offset;
-    for entry in &entries {
-        let mut header = Vec::new();
-        push_u32(&mut header, 0x02014b50);
-        push_u16(&mut header, 45);
-        push_u16(&mut header, 45);
-        push_u16(&mut header, 0x0008);
-        push_u16(&mut header, 0);
-        push_u16(&mut header, 0);
-        push_u16(&mut header, 0);
-        push_u32(&mut header, entry.crc);
-        push_u32(&mut header, u32::MAX);
-        push_u32(&mut header, u32::MAX);
-        push_u16(&mut header, entry.name.len() as u16);
-        push_u16(&mut header, 28);
-        push_u16(&mut header, 0);
-        push_u16(&mut header, 0);
-        push_u16(&mut header, 0);
-        push_u32(&mut header, 0);
-        push_u32(&mut header, u32::MAX);
-        header.extend_from_slice(&entry.name);
-        push_u16(&mut header, 0x0001);
-        push_u16(&mut header, 24);
-        push_u64(&mut header, entry.size);
-        push_u64(&mut header, entry.size);
-        push_u64(&mut header, entry.offset);
-        writer.write_all(&header).await?;
-        offset += header.len() as u64;
-    }
-    let central_size = offset - central_offset;
-    let zip64_offset = offset;
-    let mut ending = Vec::new();
-    push_u32(&mut ending, 0x06064b50);
-    push_u64(&mut ending, 44);
-    push_u16(&mut ending, 45);
-    push_u16(&mut ending, 45);
-    push_u32(&mut ending, 0);
-    push_u32(&mut ending, 0);
-    push_u64(&mut ending, entries.len() as u64);
-    push_u64(&mut ending, entries.len() as u64);
-    push_u64(&mut ending, central_size);
-    push_u64(&mut ending, central_offset);
-    push_u32(&mut ending, 0x07064b50);
-    push_u32(&mut ending, 0);
-    push_u64(&mut ending, zip64_offset);
-    push_u32(&mut ending, 1);
-    push_u32(&mut ending, 0x06054b50);
-    push_u16(&mut ending, 0);
-    push_u16(&mut ending, 0);
-    push_u16(&mut ending, u16::MAX);
-    push_u16(&mut ending, u16::MAX);
-    push_u32(&mut ending, u32::MAX);
-    push_u32(&mut ending, u32::MAX);
-    push_u16(&mut ending, 0);
-    writer.write_all(&ending).await?;
-    writer.shutdown().await?;
-    Ok(())
-}
-
-fn push_u16(output: &mut Vec<u8>, value: u16) {
-    output.extend_from_slice(&value.to_le_bytes());
-}
-fn push_u32(output: &mut Vec<u8>, value: u32) {
-    output.extend_from_slice(&value.to_le_bytes());
-}
-fn push_u64(output: &mut Vec<u8>, value: u64) {
-    output.extend_from_slice(&value.to_le_bytes());
 }
 
 async fn rtsp_status(State(state): State<AppState>) -> impl IntoResponse {
