+++ media_worker/src/media_worker.cpp
@@ -8,12 +8,8 @@
 #include <iostream>
 #include <thread>
 #include <utility>
-#include <vector>
 
 #include <unistd.h>
-
-#include "rk_mpi_sys.h"
-#include "sample_comm.h"
 
 namespace media_worker {
 
@@ -39,45 +35,6 @@ bool MediaWorker::PreflightOutputs(std::string* error) {
     if (config_.audio.enabled && !config_.audio.output_path.empty() &&
         !check("audio", config_.audio.output_path)) return false;
     return true;
-}
-
-bool MediaWorker::InitSharedRuntime(std::string* error) {
-    events_.Emit("BootProgress", {{"stage", "isp_initializing"}});
-    RK_S32 result = SAMPLE_COMM_ISP_Init(config_.isp.camera_id, RK_AIQ_WORKING_MODE_NORMAL,
-                                         RK_FALSE, config_.isp.iq_dir.c_str());
-    if (result != RK_SUCCESS) {
-        *error = "SAMPLE_COMM_ISP_Init failed: " + std::to_string(result);
-        return false;
-    }
-    isp_initialized_ = true;
-    result = SAMPLE_COMM_ISP_Run(config_.isp.camera_id);
-    if (result != RK_SUCCESS) {
-        *error = "SAMPLE_COMM_ISP_Run failed: " + std::to_string(result);
-        return false;
-    }
-    isp_running_ = true;
-    events_.Emit("BootProgress", {{"stage", "isp_ready"}});
-
-    result = RK_MPI_SYS_Init();
-    if (result != RK_SUCCESS) {
-        *error = "RK_MPI_SYS_Init failed: " + std::to_string(result);
-        return false;
-    }
-    mpi_initialized_ = true;
-    events_.Emit("BootProgress", {{"stage", "mpi_ready"}});
-    return true;
-}
-
-void MediaWorker::DeinitSharedRuntime() {
-    if (mpi_initialized_) {
-        RK_MPI_SYS_Exit();
-        mpi_initialized_ = false;
-    }
-    if (isp_running_ || isp_initialized_) {
-        SAMPLE_COMM_ISP_Stop(config_.isp.camera_id);
-        isp_running_ = false;
-        isp_initialized_ = false;
-    }
 }
 
 void MediaWorker::RequestFatalStop(const std::string& message) {
@@ -146,7 +103,8 @@ int MediaWorker::Run(const std::function<bool()>& external_stop_requested) {
                                   {"exit_code", static_cast<int>(ExitCode::kInitializationError)}});
         return static_cast<int>(ExitCode::kInitializationError);
     }
-    if (!InitSharedRuntime(&error)) {
+    runtime_ = std::make_unique<MediaRuntime>(config_.isp, &events_);
+    if (!runtime_->Init(&error)) {
         events_.Emit("FatalError", {{"stage", "shared_runtime"}, {"message", error}});
         Shutdown();
         events_.Emit("Stopped", {{"reason", "initialization_error"},
@@ -178,9 +136,9 @@ int MediaWorker::Run(const std::function<bool()>& external_stop_requested) {
     }
 
     if (config_.ai_input.control_fd >= 0) {
-        control_ = std::make_unique<MediaControl>(
-            config_.ai_input.control_fd,
-            [this](const nlohmann::json& request) { return HandleControl(request); });
+        control_service_ = std::make_unique<MediaControlService>(&config_, video_.get());
+        control_ = std::make_unique<MediaControl>(config_.ai_input.control_fd,
+            [this](const nlohmann::json& request) { return control_service_->Handle(request); });
         if (!control_->Start()) {
             events_.Emit("Warning", {{"media", "control"},
                                       {"reason", "control_start_failed"}});
@@ -236,6 +194,7 @@ void MediaWorker::Shutdown() {
         control_->Stop();
         control_.reset();
     }
+    control_service_.reset();
     if (audio_) audio_->Stop();
     if (video_) video_->Stop();
     if (audio_) {
@@ -246,94 +205,10 @@ void MediaWorker::Shutdown() {
         video_->Deinit();
         video_.reset();
     }
-    DeinitSharedRuntime();
-}
-
-nlohmann::json MediaWorker::HandleControl(const nlohmann::json& request) {
-    const std::string request_id = request.value("request_id", "");
-    auto response = nlohmann::json{{"version", 1},
-                                   {"request_id", request_id},
-                                   {"type", "ack"}};
-    if (request.value("version", 0) != 1 || request_id.empty()) {
-        return {{"version", 1},
-                {"request_id", request_id},
-                {"type", "error"},
-                {"error", "invalid control envelope"}};
+    if (runtime_) {
+        runtime_->Deinit();
+        runtime_.reset();
     }
-    const std::string command = request.value("command", "");
-    std::string error;
-    if (command == "pause_ai_frames") {
-        if (!video_->PauseAiFrames(&error)) response["type"] = "error";
-    } else if (command == "resume_ai_frames") {
-        if (!video_->ResumeAiFrames(&error)) response["type"] = "error";
-    } else if (command == "configure_ai_channel") {
-        const auto& value = request.at("ai_input");
-        AiInputConfig next = config_.ai_input;
-        next.enabled = value.value("enabled", next.enabled);
-        next.channel_id = value.value("channel_id", next.channel_id);
-        next.width = value.value("width", next.width);
-        next.height = value.value("height", next.height);
-        next.fps = value.value("fps", next.fps);
-        next.pixel_format = value.value("pixel_format", next.pixel_format);
-        next.fit_mode = value.value("fit_mode", next.fit_mode);
-        next.buffer_count = value.value("buffer_count", next.buffer_count);
-        next.depth = value.value("depth", next.depth);
-        WorkerConfig candidate = config_;
-        candidate.ai_input = next;
-        const auto errors = ValidateConfig(candidate);
-        if (!errors.empty()) {
-            response["type"] = "error";
-            response["error"] = errors;
-            return response;
-        }
-        if (!video_->ReconfigureAiInput(next, &error)) {
-            response["type"] = "error";
-        } else {
-            config_.ai_input = next;
-            response["event"] = "ai_input_ready";
-            response["ai_input"] = value;
-        }
-    } else if (command == "probe_region_capability") {
-        const auto capability = video_->ProbeRegionCapability(&error);
-        if (!error.empty()) response["type"] = "error";
-        response["event"] = "region_capability";
-        for (auto item = capability.begin(); item != capability.end(); ++item) {
-            response[item.key()] = item.value();
-        }
-    } else if (command == "set_osd_mode") {
-        const std::string mode = request.value("mode", "off");
-        if (!video_->SetOsdMode(mode, &error)) {
-            response["type"] = "error";
-        } else {
-            response["mode"] = mode;
-        }
-    } else if (command == "update_regions") {
-        const int main_width = request.value("main_width", 0);
-        const int main_height = request.value("main_height", 0);
-        if (main_width != config_.video.width || main_height != config_.video.height) {
-            response["type"] = "error";
-            error = "region coordinate space does not match the main video";
-        } else {
-            std::vector<OsdRegion> regions;
-            for (const auto& value : request.value("regions", nlohmann::json::array())) {
-                if (regions.size() >= 32) break;
-                OsdRegion region{value.value("x", 0), value.value("y", 0),
-                                 value.value("width", 0), value.value("height", 0)};
-                if (region.width < 2 || region.height < 2) continue;
-                regions.push_back(region);
-            }
-            if (!video_->UpdateRegions(regions, request.value("ttl_ms", 300), &error)) {
-                response["type"] = "error";
-            } else {
-                response["regions"] = regions.size();
-            }
-        }
-    } else {
-        response["type"] = "error";
-        error = "unknown media control command";
-    }
-    if (!error.empty()) response["error"] = error;
-    return response;
 }
 
 }  // namespace media_worker
