+++ media_worker/src/media_worker.h
@@ -11,6 +11,8 @@
 #include "event_emitter.h"
 #include "metrics_sampler.h"
 #include "media_control.h"
+#include "media_control_service.h"
+#include "media_runtime.h"
 #include "video_pipeline.h"
 
 namespace media_worker {
@@ -31,12 +33,9 @@ public:
 
 private:
     bool PreflightOutputs(std::string* error);
-    bool InitSharedRuntime(std::string* error);
-    void DeinitSharedRuntime();
     void RequestFatalStop(const std::string& message);
     std::string RuntimeError() const;
     void EmitMetrics(double elapsed_seconds);
-    nlohmann::json HandleControl(const nlohmann::json& request);
     void Shutdown();
 
     WorkerConfig config_;
@@ -44,12 +43,11 @@ private:
     std::unique_ptr<VideoPipeline> video_;
     std::unique_ptr<AudioPipeline> audio_;
     std::unique_ptr<MediaControl> control_;
+    std::unique_ptr<MediaControlService> control_service_;
+    std::unique_ptr<MediaRuntime> runtime_;
     std::atomic<bool> stop_requested_{false};
     mutable std::mutex error_mutex_;
     std::string runtime_error_;
-    bool isp_initialized_ = false;
-    bool isp_running_ = false;
-    bool mpi_initialized_ = false;
     MetricsSampler video_metrics_sampler_;
     MetricsSampler audio_metrics_sampler_;
 };
