+++ scripts/build-rv1106.sh
@@ -12,6 +12,7 @@ CROSS_GCC="${TOOLCHAIN_DIR}/bin/${CROSS_PREFIX}-gcc"
 CROSS_GXX="${TOOLCHAIN_DIR}/bin/${CROSS_PREFIX}-g++"
 CROSS_AR="${TOOLCHAIN_DIR}/bin/${CROSS_PREFIX}-ar"
 OPENSSL_DIR="${AIPC_OPENSSL_DIR:-${AIPC_SDK_ROOT}/sysdrv/source/buildroot/buildroot-2023.02.6/output/host/arm-buildroot-linux-uclibcgnueabihf/sysroot/usr}"
+NATIVE_INSTALL="${PROJECT_ROOT}/target/native/RV1106Release/install"
 
 for tool in "${CROSS_GCC}" "${CROSS_GXX}" "${CROSS_AR}"; do
     if [ ! -x "${tool}" ]; then
@@ -19,8 +20,6 @@ for tool in "${CROSS_GCC}" "${CROSS_GXX}" "${CROSS_AR}"; do
         exit 1
     fi
 done
-
-"${SCRIPT_DIR}/fetch-ai-deps.sh"
 
 if [ ! -f "${OPENSSL_DIR}/include/openssl/ssl.h" ] || [ ! -e "${OPENSSL_DIR}/lib/libssl.so" ]; then
     echo "missing target OpenSSL 1.1.1 sysroot at ${OPENSSL_DIR}; set AIPC_OPENSSL_DIR" >&2
@@ -30,6 +29,30 @@ fi
 rustup component add rust-src --toolchain nightly >/dev/null
 
 export AIPC_SDK_ROOT
+
+NATIVE_CMAKE_ARGS=()
+for dependency in NLOHMANN_JSON LUA VISIONG LUCKFOX_RKMPI; do
+    override="FETCHCONTENT_SOURCE_DIR_${dependency}"
+    if [ -n "${!override:-}" ]; then
+        NATIVE_CMAKE_ARGS+=("-D${override}=${!override}")
+    fi
+done
+if [ "${AIPC_FETCHCONTENT_OFFLINE:-OFF}" = "ON" ]; then
+    NATIVE_CMAKE_ARGS+=("-DAIPC_FETCHCONTENT_OFFLINE=ON")
+fi
+
+cd "${PROJECT_ROOT}/native"
+cmake --preset RV1106Release "${NATIVE_CMAKE_ARGS[@]}"
+cmake --build --preset RV1106Release
+cmake --install "${PROJECT_ROOT}/target/native/RV1106Release/build"
+
+for native_binary in media_worker ai_worker; do
+    if [ ! -x "${NATIVE_INSTALL}/bin/${native_binary}" ]; then
+        echo "native build did not install ${native_binary}" >&2
+        exit 1
+    fi
+done
+
 export CC_armv7_unknown_linux_uclibceabihf="${CROSS_GCC}"
 export CXX_armv7_unknown_linux_uclibceabihf="${CROSS_GXX}"
 export AR_armv7_unknown_linux_uclibceabihf="${CROSS_AR}"
@@ -47,5 +70,11 @@ cargo +nightly build \
     --release \
     --package aipc-daemon
 
+install -m 0755 "${NATIVE_INSTALL}/bin/media_worker" \
+    "target/${TARGET}/release/media_worker"
+install -m 0755 "${NATIVE_INSTALL}/bin/ai_worker" \
+    "target/${TARGET}/release/ai_worker"
+
 file "target/${TARGET}/release/aipc-daemon"
 file "target/${TARGET}/release/media_worker"
+file "target/${TARGET}/release/ai_worker"
